refactor: move community tabs into components
