+++ src/components/community/cards/HotSearchesCard.tsx
