+++ src/components/community/cards/KpiCard.tsx
