+++ src/components/community/cards/MapCard.tsx
