+++ src/components/community/cards/TrendCard.tsx
