+++ src/pages/messages/MessagesPage.tsx
@@ -1,6 +1,5 @@
+import { MessagesTab } from '@components/community/MessagesTab'
 import { BaseLayout } from '@components/layout/BaseLayout/BaseLayout'
-
-import { MessagesTab } from '@src/pages/community/tabs/MessagesTab'
 
 export const MessagesPage = () => (
   <BaseLayout>
