+++ src/pages/stats/StatsPage.tsx
@@ -1,6 +1,5 @@
+import { StatsTab } from '@components/community/StatsTab'
 import { BaseLayout } from '@components/layout/BaseLayout/BaseLayout'
-
-import { StatsTab } from '@src/pages/community/tabs/StatsTab'
 
 export const StatsPage = () => (
   <BaseLayout>
